server/requirements.txt
asgiref==3.7.2
Django==4.2.10
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
pillow==9.0.0
psycopg2==2.9.9
django-cors-headers==4.3.1